Pk id overwrite bug (#4)
* Primary Key maintenance issues

A few things regarding keys and entities were changed to try and be
more in line with FHIR actually works. In a nutshell, FHIR Elements are
limited to within the context of their Resource, and should not be
shared across resources (e.g. 2 Patients should not reference the same
HumanName). The current architecture and Realm limitations prevent me -
at least at this point until I drink a boatload more coffee - from
being 100% in spirit with this. So the following has been done:

- If an Element of a resource has already been saved, then populating
that element from JSON will now overwrite value fields, but maintain
the `pk` field. In essence, updating that item.

- RealmLists of Elements are now completely blown away and re-populated
when deserializing from JSON.

- I’ve yet to find a good way to copy-on-write in Realm objects when
assigning element values to resources (without losing query
functionality, or notification functionality, or basically all Realm
functionality). As a result, it is up to the user to ensure they copy
before writing. This sucks, but to help in this all Resources and
Elements now implement NSCopy, which does a deep copy but ignores `pk`
fields. eg. `somePatient.name.copy() as! HumanName` will return a deep
copy of `somePatient`’s name, but as brand new realm objects.

* updated readme, added tests, upset functions