Cascade deletes (#13)
* cascade-deletes: CascadeDeletable protocol

Created a `CascadeDeletable` protocol and applied default
implementations to `FHIRAbstractBase` and `RealmSwift.List`. The
`FHIRAbstractBase` will go through and delete each element in each
child list, as well as each child object. The `RealmSwift.List`
implementation simply iterates through each element and deletes it.

In both cases, if the thing being deleted is a `CascadeDeletable`, then
`cascadeDelete()` will be called. Otherwise it will delete it using
whatever `realm` is associated with that object. If the object is
unmanaged then the object, naturally, will not be deleted.

* cascade-deletes: wordsmithing

* cascade-deletes: Added new template files for when everything is generated.